app/pyproject.toml
[project]
name = "toconcursos-app"
version = "0.1.0"
description = "TOConcursos backend app"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "pydantic",
    "SQLAlchemy",
    "python-dotenv",
    "alembic",
    "pytest",
    "pytest-asyncio",
    "httpx",
    "passlib[bcrypt]",
    "python-jose",
    "psycopg2-binary",
    "minio>=7.2.18",
    "python-multipart>=0.0.20",
    "google-genai",
    "google-auth-oauthlib>=1.2.3",
    "google-api-python-client>=2.187.0",
]
